Milkomeda-C1 Testnet network has been added to the supported chains

+++ contracts/OracleFacingGeoConsumer.sol
@@ -4,5 +4,5 @@ pragma solidity ^0.8.7;
 import "@shambadynamic/contracts/contracts/ShambaGeoConsumer.sol";
 
 contract OracleFacingGeoConsumer is ShambaGeoConsumer {
-    constructor(uint256 _chainId) ShambaGeoConsumer(_chainId) {}
+    constructor(uint64 _chainId) ShambaGeoConsumer(_chainId) {}
 }
